feat : 코치님 피드백 반영 - ConcertFacade가 무거워 질 수 있어, 분리가 필요.
