+++ src/main/java/hhplus/concertreservationservice/presentation/concert/schedular/ConcertSchedular.java
@@ -2,6 +2,7 @@ package hhplus.concertreservationservice.presentation.concert.schedular;
 
 
 import hhplus.concertreservationservice.application.concert.facade.ConcertFacade;
+import hhplus.concertreservationservice.application.concert.facade.ConcertReservationFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -10,13 +11,13 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class ConcertSchedular {
 
-    private final ConcertFacade concertFacade;
+    private final ConcertReservationFacade concertReservationFacade;
 
     // 1분마다
     // 예약 만료 스케줄러
     @Scheduled(cron = "0 */1 * * * *")
     public void expireReservationProcess(){
-        concertFacade.expireReservationProcess();
+        concertReservationFacade.expireReservationProcess();
 
     }
